Added some changes to prevent Fatal Error
Added some changes to prevent Fatal Error exception being called when the client could not instantiate.
This because this kind exception is not catchable in any way. It stops the execution of the program, without the possibility for example, to perform some log activity.
The best way to perform the client creation will be, probably, to launch the creation without the possibility of some fatal error, and to thow, instead another kind of catchable exception to the user.